+++ changelog.md
@@ -25,6 +25,10 @@
 ### Changes
 - Tools that can place torches now place crystal torches
 	- These torches do not drop items when broken
+- Crafting a Crystal AIOT now gives you the skill points from the tools used to make it
+
+### Fixes
+- Fixes other items being able to be added in to the crystal elytra recipe
 
 ----------------------- 1.2.2 Changes -----------------------
 ### Fixes